Fix compilation

+++ osmflat/examples/render-roads.rs
@@ -23,7 +23,7 @@ struct GeoCoord {
 
 /// Convert osmflat Node into GeoCoord.
 impl GeoCoord {
-    fn from_node(node: &Node, coord_scale: u64) -> Self {
+    fn from_node(node: &Node, coord_scale: i32) -> Self {
         Self {
             lat: node.lat() as f64 / coord_scale as f64,
             lon: node.lon() as f64 / coord_scale as f64,
